refactored test cases for numeric reps
